requirements.txt
watchdog==6.0.0
requests~=2.32.3
PyYAML~=6.0.2